Fix type ElementCount in Model

+++ TestXmlAnalytics/Model/FileAnalyzeModel.cs
@@ -2,7 +2,7 @@
 {
     public class FileAnalyzeModel
     {
-        public int ElementCount { get; set; }
+        public long ElementCount { get; set; }
         public IEnumerable<string> ElementNames { get; set; } = [];
         public Dictionary<string, long> AttributeSum { get; set; } = [];
     }
